# lib/api-zod/src/generated/types/createMonthlyActualBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMonthlyActualBody {
  budgetLineId: number;
  month: number;
  year: number;
  actualAmount: number;
  invoiceRef?: string;
}