# lib/api-zod/src/generated/types/createMonthlyPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMonthlyPlanBody {
  budgetLineId: number;
  month: number;
  year: number;
  plannedAmount: number;
  boardAmount?: number;
}