# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalBudget: number;
  spentYtd: number;
  remaining: number;
  fixedRunRate: number;
  activeAlerts: number;
  budgetUtilisation: number;
  monthsElapsed: number;
  totalMonths: number;
}