# lib/api-zod/src/generated/types/marketingEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketingEvent {
  id: number;
  name: string;
  /** @nullable */
  eventDate?: Date | null;
  status: string;
  /** @nullable */
  estimatedCost?: number | null;
  /** @nullable */
  budgetLineId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}